Project structure reorganized
- Editor config added to the project
- Components organized in to folders
- Pages organized in to folders
