+++ android/app/src/main/assets/capacitor.config.json
@@ -1,0 +1,6 @@
+{
+	"appId": "io.ionic.starter",
+	"appName": "ForAll",
+	"webDir": "build",
+	"bundledWebRuntime": false
+}
