+++ binaries/ForALL_devTools.apk
