+++ local.properties
@@ -5,4 +5,5 @@
 # For customization when using a Version Control System, please read the
 # header note.
 #Mon Oct 04 13:50:49 BRT 2021
+
 sdk.dir=C\:\\Users\\alexa\\AppData\\Local\\Android\\Sdk
