+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ForAll",
+  "name": "for-all",
   "version": "0.0.1",
   "private": true,
   "dependencies": {
