+++ src/components/AudioComp.css
@@ -1,13 +1,13 @@
 #startButton {
-  margin: 10px;
-  width: 50px;
-  height: 50px;
-  background-color: rgb(0, 255, 66);
+	margin: 10px;
+	width: 50px;
+	height: 50px;
+	background-color: rgb(0, 255, 66);
 }
 
 .frequencyBands {
-  padding: 6px;
-  margin: 2px;
-  transform: rotatex(180deg);
-  transform-origin: top;
+	padding: 6px;
+	margin: 2px;
+	transform: rotatex(180deg);
+	transform-origin: top;
 }
