+++ src/components/AudioViz/AudioVizInterface.ts
@@ -1,0 +1,5 @@
+interface AudioVizProps {
+	vizFlag: number;
+}
+
+export default AudioVizProps;
