+++ src/components/ExploreContainer/ExploreContainerInterface.ts
@@ -1,0 +1,5 @@
+interface ContainerProps {
+	name: string;
+}
+
+export default ContainerProps;
