+++ src/components/MyHeader/MyHeaderInterface.ts
@@ -1,0 +1,5 @@
+interface MyHeaderProps {
+	name: String;
+}
+
+export default MyHeaderProps
